Fix function dircheck
Change the return value of function dircheck from void to int. Use
return 0 and 1 instead of exit(EXIT_SUCCESS) or exit(EXIT_FAILURE). The
old version will exit whatever the input is a directory or not.
Also, make the command line more strict. That is, must input the local
directory in the end of the command line.